requirements.txt
streamlit
pandas
altair
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib